linux26-proc-interface-use-long-not-int-20040511
FIXES 4462

arguments to sys_call are long, not int
ioctl returns -ERROR, not -1
return -EINVAL on ioctls we don't know

+++ src/config/afs_args.h
@@ -183,11 +183,11 @@ typedef struct cm_initparams_v1 {
 #define VIOC_SYSCALL _IOW(VIOC_SYSCALL_TYPE,1,void *)
  
 struct afsprocdata {
-  int param4;
-  int param3;
-  int param2;
-  int param1;
-  int syscall;
+  long param4;
+  long param3;
+  long param2;
+  long param1;
+  long syscall;
 };
  
 #endif
